+++ ignite/services/chain/lint_test.go
@@ -1,0 +1,7 @@
+package chain_test
+
+import "testing"
+
+func TestChainLint(t *testing.T) {
+
+}